s2/s2 fixes
