+++ src/services/reducers/burger.js
@@ -1,138 +1,9 @@
 import {
-  GET_INGREDIENTS_FAILED,
-  GET_INGREDIENTS_REQUEST,
-  GET_INGREDIENTS_SUCCESS,
-  SET_CURRENT_TAB,
   ADD_SELECTED_ITEM,
   MOVE_SELECTED_ITEM,
   DELETE_SELECTED_ITEM,
   COUNT_TOTAL,
-  DELETE_CURRENT_INGREDIENT,
-  SET_CURRENT_INGREDIENT,
-  GET_ORDER_FAILED,
-  GET_ORDER_REQUEST,
-  GET_ORDER_SUCCESS,
-  ORDER_OPEN,
-  ORDER_CLOSE,
-  DETAILS_OPEN,
-  DETAILS_CLOSE,
-  INCREASE_INGREDIENT_COUNT,
-  DECREASE_INGREDIENT_COUNT,
-  SET_INGREDIENT_COUNT,
-  SET_INGREDIENT_COUNT_BY_TYPE,
-  DROP_INGRIDIENTS_STATE,
-  DROP_ORDER_STATE
 } from '../actions/burger';
-
-const ingredientsInitialState = {
-  all: [],
-  ingredientsRequest: false,
-  ingredientsFailed: false,
-  currentTab: 'buns'
-};
-
-export const ingredients = (state = ingredientsInitialState, action) => {
-  switch (action.type) {
-    case GET_INGREDIENTS_REQUEST: {
-      return {
-        ...state,
-        ingredientsRequest: true
-      }
-    }
-    case GET_INGREDIENTS_SUCCESS: {
-      return {
-        ...state,
-        ingredientsRequest: false,
-        ingredientsFailed: false,
-        all: action.ingredients
-      }
-    }
-    case GET_INGREDIENTS_FAILED: {
-      return {
-        ...state,
-        ingredientsRequest: false,
-        ingredientsFailed: true
-      }
-    }
-
-    case SET_CURRENT_TAB: {
-      return {
-        ...state,
-        currentTab: action.tab
-      }
-    }
-    case INCREASE_INGREDIENT_COUNT: {
-      return {
-        ...state,
-        all: [...state.all].map(item => item._id === action.id ? { ...item, __v: item.__v + 1 } : item)
-      }
-    }
-    case DECREASE_INGREDIENT_COUNT: {
-      return {
-        ...state,
-        all: [...state.all].map(item => item._id === action.id ? { ...item, __v: item.__v - 1 } : item)
-      }
-    }
-    case SET_INGREDIENT_COUNT: {
-      return {
-        ...state,
-        all: [...state.all].map(item => item._id === action.id ? { ...item, __v: action.count } : item)
-      }
-    }
-    case SET_INGREDIENT_COUNT_BY_TYPE: {
-      return {
-        ...state,
-        all: [...state.all].map(item => item.type === action.ingredientType ? { ...item, __v: action.count } : item)
-      }
-    }
-    case DROP_INGRIDIENTS_STATE: {
-      return {
-        ...state,
-        all: []
-      }
-    }
-    // case : {return{...state}}
-    default: {
-      return state;
-    }
-  }
-}
-
-const initialDetailsState = {
-  currentIngredient: {},
-  isDetailsModalOpen: false
-}
-
-export const details = (state = initialDetailsState, action) => {
-  switch (action.type) {
-    case DETAILS_OPEN: {
-      return {
-        ...state,
-        isDetailsModalOpen: true
-      }
-    }
-    case DETAILS_CLOSE: {
-      return {
-        ...state,
-        isDetailsModalOpen: false
-      }
-    }
-    case SET_CURRENT_INGREDIENT: {
-      return {
-        ...state,
-        currentIngredient: action.item
-      }
-    }
-    case DELETE_CURRENT_INGREDIENT: {
-      return {
-        ...state,
-        currentIngredient: {}
-      }
-    }
-    default: { return state; }
-  }
-
-}
 
 function countTotal(content) {
   let total = 0;
@@ -191,62 +62,6 @@ export const burger = (state = burgerInitialState, action) => {
         selected: { bun: state.selected.bun, filling: deleteElement(state.selected.filling, action.id) }
       }
     }
-    // case : {return{...state}}
-    default: {
-      return state;
-    }
-  }
-}
-
-const orderInitialState = {
-  order: {},
-  orderRequest: false,
-  orderFailed: false,
-  isOrderModalOpen: false
-};
-
-export const order = (state = orderInitialState, action) => {
-  switch (action.type) {
-    case GET_ORDER_REQUEST: {
-      return {
-        ...state,
-        orderRequest: true
-      }
-    }
-    case GET_ORDER_SUCCESS: {
-      return {
-        ...state,
-        orderRequest: false,
-        orderFailed: false,
-        order: action.order
-      }
-    }
-    case GET_ORDER_FAILED: {
-      return {
-        ...state,
-        orderRequest: false,
-        orderFailed: true
-      }
-    }
-    case ORDER_OPEN: {
-      return {
-        ...state,
-        isOrderModalOpen: true
-      }
-    }
-    case ORDER_CLOSE: {
-      return {
-        ...state,
-        isOrderModalOpen: false
-      }
-    }
-    case DROP_ORDER_STATE: {
-      return {
-        ...state,
-        order: {}
-      }
-    }
-    // case : {return{...state}}
     default: {
       return state;
     }
