+++ src/services/reducers/index.js
@@ -1,4 +1,7 @@
 import { combineReducers } from 'redux';
-import { ingredients, details, burger, order } from './burger';
+import { ingredients } from './ingredients';
+import { details } from './details';
+import { burger } from './burger';
+import { order } from './order';
 
 export const rootReducer = combineReducers({ ingredients, details, burger, order})
