+++ src/services/reducers/selectors.js
@@ -1,0 +1,4 @@
+export const selectIngredients = (state) => state.ingredients;
+export const selectDetails = (state) => state.details;
+export const selectBurger = (state) => state.burger;
+export const selectOrder = (state) => state.order;
